+++ CHANGELOG.md
@@ -8,16 +8,18 @@ If you like this project and find it useful, please consider giving it a star on
 
 - Unless you are using docker (in that case all is already updated when you pull the image), please update Matterbridge to 1.5.4 to work with matterbridge-shelly >= 0.9.5.  This is a one time issue due to the update to matter.js 0.10.0.
 
-## [0.9.8] - 2024-09-10
+## [0.9.8] - 2024-09-12
 
 ### Added
 
 ### Verified
 
 - [shelly]: Verified shellyplus010v with firmware v. 1.4.2 and added Jest test.
+- [shelly]: Verified shellyplusi4 (DC) with firmware v. 1.4.2 and added Jest test.
 
 ### Changed
 
+- [package]: Updated typescript to 5.6.2.
 - [package]: Updated dependencies.
 
 <a href="https://www.buymeacoffee.com/luligugithub">